Update installer to use installed directories
It occured to me that we can simplify the installer greatly by running the
install command for each binary set, and just doing a recursive file copy in the
installer for x64 and x32.  This is adventageous as it lets us package whatever
configuration we built, and will include any files created as part of the
install in the future.  It also makes the installer script significantly smaller

Reviewed-by: Tomas Mraz <tomas@openssl.org>
Reviewed-by: Matt Caswell <matt@openssl.org>
(Merged from https://github.com/openssl/openssl/pull/4)